+++ ArchipelagoDeathCounter/Program.cs
@@ -3,6 +3,8 @@ using System.Text;
 using Archipelago.MultiClient.Net.Enums;
 using Archipelago.MultiClient.Net.Models;
 using ArchipelagoDeathCounter;
+using ArchipelagoDeathCounter.LoggerConsole;
+using ArchipelagoDeathCounter.LoggerConsole.Commands;
 using Backbone;
 using ImGuiNET;
 using Raylib_cs;
@@ -15,11 +17,14 @@ ui.Loop();
 class UserInterface() : Backbone.Backbone("DeathLinkipelago", 650, 700)
 {
     public int Counter = 0;
-    private ArchipelagoClient Client = new();
-    private long LastUpdate = DateTimeOffset.Now.ToUnixTimeMilliseconds();
+    public static ArchipelagoClient Client = new();
+    public long LastUpdate = DateTimeOffset.Now.ToUnixTimeMilliseconds();
 
     public override void Init()
     {
+        Logger.Initialize();
+        Logger.Log("init");
+        CommandRegister.RegisterCommandFile(typeof(DefaultCommands));
         Raylib.SetTargetFPS(30);
         Client.Init();
     }
@@ -34,7 +39,37 @@ class UserInterface() : Backbone.Backbone("DeathLinkipelago", 650, 700)
 
     public override void Render()
     {
-        Client.Render();
+        if (ImGui.BeginTabBar("tabs"))
+        {
+
+            if (ImGui.BeginTabItem("Deathlinkipelago"))
+            {
+                Client.Render();
+                ImGui.EndTabItem();
+            }
+
+            if (ImGui.BeginTabItem("Text Client"))
+            {
+                ImGui.Text("Coming Soon:tm:");
+                ImGui.Text("maybe v.0.9");
+                ImGui.EndTabItem();
+            }
+            
+            if (ImGui.BeginTabItem("Hints"))
+            {
+                ImGui.Text("Coming Soon:tm:");
+                ImGui.Text("maybe v.0.9");
+                ImGui.EndTabItem();
+            }
+
+            if (ImGui.BeginTabItem("Console"))
+            {
+                GameConsole.Render();
+                ImGui.EndTabItem();
+            }
+            
+            ImGui.EndTabBar();
+        }
     }
 }
 
@@ -51,14 +86,14 @@ public class ArchipelagoClient
     public string[]? Error;
     public string[] PlayerNames = [];
     public string LastPersonToBlame = "N/A";
-    public float SaveCooldown = 3;
+    public float SaveCooldown = 1.5f;
     public float DeathCooldown = 30;
     public float TimeSinceLastDeath;
     public float LongestTimeSinceLastDeath;
     public bool HasChangedSinceSave;
     public bool HasDeathButton;
     public bool SendTrapsAfterGoal;
-    
+
     public Dictionary<string, int> Deaths = [];
 
     public void Init() { Connection.LoadInitData(); }
@@ -70,7 +105,7 @@ public class ArchipelagoClient
         Connection.Update();
         TimeSinceLastDeath += deltaTime;
 
-        if (DeathCooldown < 3)
+        if (DeathCooldown < 1.5)
         {
             DeathCooldown += deltaTime;
         }
@@ -86,8 +121,6 @@ public class ArchipelagoClient
                 {
                     Connection.SendDeathLink("The Will of God");
                 }
-
-                AddDeath(Connection.Slot);
             }
 
             Connection.UsedItems["Death Trap"]++;
@@ -141,12 +174,12 @@ public class ArchipelagoClient
                 RenderItemTable();
                 ImGui.NewLine();
                 RenderDeathTable();
-                
+
                 ImGui.Text($"Time since last death: [{GetTime(TimeSinceLastDeath)}]");
                 ImGui.Text(
                     $"Longest Time since last death: [{GetTime(Math.Max(LongestTimeSinceLastDeath, TimeSinceLastDeath))}]");
                 ImGui.Text($"Last recorded death link was from: [{LastPersonToBlame}]");
-                
+
                 ImGui.NewLine();
                 RenderShopTable();
 
@@ -230,6 +263,7 @@ public class ArchipelagoClient
 
     public void RenderShopTable()
     {
+        var canAfford = Connection.GetItemAmount("Death Coin") > 0;
         ImGui.Text($"Shop Level: [{Connection.HeldItems["Progressive Death Shop"]}]");
         if (ImGui.BeginTable("Shop Table", 4, TableFlags | ImGuiTableFlags.SizingFixedFit))
         {
@@ -245,14 +279,22 @@ public class ArchipelagoClient
                 ImGui.TableSetColumnIndex(0);
                 ImGui.Text($"Item {id + 1} ");
                 ImGui.TableNextColumn();
-                ImGui.TextColored(GetItemColor(location), $"{location.ItemName} ".Replace("Trap", "Sheild"));
+                ImGui.TextColored(GetItemColor(location),
+                    $"{location.ItemName} ".Replace(" Trap", " Sheild").Replace(" Shield", " Tarp"));
                 ImGui.TableNextColumn();
                 ImGui.Text($"{location.Player.Name} ");
                 ImGui.TableNextColumn();
 
-                if (ImGui.Button($"Buy Item {id + 1}"))
+                if (canAfford)
                 {
-                    Connection.BuyCheck(id, this);
+                    if (ImGui.Button($"Buy Item {id + 1}"))
+                    {
+                        Connection.BuyCheck(id, this);
+                    }
+                }
+                else
+                {
+                    ImGui.TextColored(Red, "Cannot Afford");
                 }
 
                 ImGui.TableNextColumn();
@@ -261,8 +303,8 @@ public class ArchipelagoClient
 
         ImGui.EndTable();
     }
-    
-    public void AddDeath(string slot)
+
+    public void AddDeath(string slot, int amount = 1)
     {
         if (!PlayerNames.Contains(slot))
         {
@@ -271,8 +313,9 @@ public class ArchipelagoClient
 
         LastPersonToBlame = slot;
         Deaths.TryAdd(slot, 0);
-        Deaths[slot]++;
-        Connection.HeldItems["Death Coin"]++;
+        Deaths[slot] += amount;
+        Connection.HeldItems["Death Coin"] += amount;
+        OrderDeaths();
         LongestTimeSinceLastDeath = Math.Max(LongestTimeSinceLastDeath, TimeSinceLastDeath);
         TimeSinceLastDeath = 0;
         HasChangedSinceSave = true;
@@ -297,8 +340,11 @@ public class ArchipelagoClient
 
     public Vector4 GetItemColor(ScoutedItemInfo item)
     {
+        if (item.ItemName.Contains(" Shield")) return Red;
         if (item.Flags.HasFlag(ItemFlags.Advancement)) return Gold;
         if (item.Flags.HasFlag(ItemFlags.Trap) || item.Flags.HasFlag(ItemFlags.NeverExclude)) return Blue;
         return SkyBlue;
     }
+
+    public void OrderDeaths() => Deaths = Deaths.OrderBy(kv => kv.Value).ToDictionary(kv => kv.Key, kv => kv.Value);
 }